neo/io/plexonio.py
line 141, 142
nb_samples = np.zeros(....., dtype=int)
sample_positions = np.zeros(....., dtype=int)

+++ NeoAnalysis_Py3.5/NeoAnalysis/neo/io/plexonio.py
@@ -138,8 +138,8 @@ class PlexonIO(BaseIO):
 
         ## Step 2 : a first loop for counting size
         # signal
-        nb_samples = np.zeros(len(slowChannelHeaders))
-        sample_positions = np.zeros(len(slowChannelHeaders))
+        nb_samples = np.zeros(len(slowChannelHeaders),dtype=int)
+        sample_positions = np.zeros(len(slowChannelHeaders),dtype=int)
         t_starts = np.zeros(len(slowChannelHeaders), dtype='f')
 
         #spiketimes and waveform
